+++ website2.js
@@ -51,34 +51,46 @@ let lists = defaultLists;
 // //Event Listener Functions
 import * as eventListeners from "./eventListeners2.js";
 
+//Functions
+import {
+createTask, 
+hideTaskDetails, 
+resetPriorityRadioButtons, 
+toggleInputField, 
+isValidDate, 
+saveAndRender, 
+render, 
+} from './functions2.js';
+
 // ^^^^^Global Declarations^^^^^  moved into globalDeclarations.js
 // ^^^^^Local Storage Elements^^^^^  moved into localStorage.js
 
 //Event Listeners to be moved to eventListeners2
-taskContainer.addEventListener("click", (e) => {
-  if (e.target.tagName.toLowerCase() === "input") {
-    const selectedList = lists.find((list) => list.id === selectedListId.value);
-    const selectedTask = selectedList.tasks.find(
-      (task) => task.id === e.target.id
-    );
-    selectedTask.complete = e.target.checked;
-    save();
-    renderTaskCount(selectedList);
-  }
-});
 
-newListForm.addEventListener("submit", (e) => {
-  e.preventDefault();
-  const listName = newListInput.value;
-  if (listName == null || listName == "") return;
-  const list = createList(listName);
-  // console.log('EL newListFormNAME:', list);
-  newListInput.value = null;
-  lists.push(list);
-  console.log("EL newListFormNAME:", list);
-  console.log("EL newListFormARRAY:", lists);
-  saveAndRender();
-});
+// taskContainer.addEventListener("click", (e) => {
+//   if (e.target.tagName.toLowerCase() === "input") {
+//     const selectedList = lists.find((list) => list.id === selectedListId.value);
+//     const selectedTask = selectedList.tasks.find(
+//       (task) => task.id === e.target.id
+//     );
+//     selectedTask.complete = e.target.checked;
+//     save();
+//     renderTaskCount(selectedList);
+//   }
+// });
+
+// newListForm.addEventListener("submit", (e) => {
+//   e.preventDefault();
+//   const listName = newListInput.value;
+//   if (listName == null || listName == "") return;
+//   const list = createList(listName);
+//   // console.log('EL newListFormNAME:', list);
+//   newListInput.value = null;
+//   lists.push(list);
+//   console.log("EL newListFormNAME:", list);
+//   console.log("EL newListFormARRAY:", lists);
+//   saveAndRender();
+// });
 
 newTaskForm.addEventListener("submit", (e) => {
   e.preventDefault();
@@ -129,15 +141,19 @@ clearCompleteTasksButton.addEventListener("click", (e) => {
 });
 
 deleteListButton.addEventListener("click", (e) => {
-  lists = lists.filter((list) => list.id !== selectedListId.value);
+const deleteList = selectedListId.value;
+  // lists = lists.filter((list) => list.id !== selectedListId.value);
+  lists = lists.filter((list) => list.id !== deleteList);
+  console.log('EL dlb:', lists);
   selectedListId.value = null;
   saveAndRender();
+  // selectedListId.value = null;
 });
 
 function setupEventListeners() {
   listsContainer.addEventListener("click", eventListeners.handleListsContainer);
-  // taskContainer.addEventListener('click', eventListeners.handleTaskContainer);
-  // newListForm.addEventListener('click', eventListeners.handleNewListForm);
+  taskContainer.addEventListener("click", eventListeners.handleTaskContainer);
+  newListForm.addEventListener('click', eventListeners.handleNewListForm);
   // newTaskForm.addEventListener('click', eventListeners.handleNewTaskForm);
   // taskNoteButton.addEventListener('click', eventListeners.handleTaskNoteButton);
   // taskPriorityButton.addEventListener('click', eventListeners.handleTaskPriorityButton);
@@ -147,149 +163,6 @@ function setupEventListeners() {
 }
 setupEventListeners();
 
-//  ^^^^^Functions^^^^^
-function createList(name) {
-  console.log("F createList:", name);
-  return {
-    id: Date.now().toString(),
-    name: name,
-    tasks: [],
-  };
-}
-
-function createTask(name, note, priority, dueDate) {
-  return {
-    id: Date.now().toString(),
-    name: name,
-    note: note || "",
-    priority: priority || "",
-    dueDate: dueDate || "",
-    complete: false,
-  };
-}
-
-function hideTaskDetails() {
-  taskNoteInput.style.display = "none";
-  taskPriorityOptions.style.display = "none";
-  taskDueDateInput.style.display = "none";
-}
-
-function resetPriorityRadioButtons(container) {
-  const radioButtons = container.querySelectorAll('input[type="radio"]');
-  radioButtons.forEach((radioButton) => {
-    radioButton.checked = false;
-  });
-}
-
-function toggleInputField(inputField) {
-  if (inputField.style.display === "block") {
-    inputField.style.display = "none";
-  } else {
-    inputField.style.display = "block";
-    inputField.focus();
-  }
-}
-
-function isValidDate(dateString) {
-  if (dateString === "") {
-    return true;
-  }
-  const datePattern = /^(0[1-9]|1[0-2])\/(0[1-9]|[12]\d|3[01])\/\d{2}$/;
-  if (!datePattern.test(dateString)) {
-    return false;
-  }
-  const [month, day, year] = dateString.split("/").map(Number);
-  if (month < 1 || month > 12) {
-    return false;
-  }
-  const daysInMonth = new Date(year, month, 0).getDate();
-  if (day < 1 || day > daysInMonth) {
-    return false;
-  }
-  if (year < 0 || year > 99) {
-    return false;
-  }
-  return true;
-}
-
-export function saveAndRender() {
-  save();
-  render();
-}
-
-//Local Storage Creation/Save
-function save() {
-  localStorage.setItem(LOCAL_STORAGE_LIST_KEY, JSON.stringify(lists));
-  localStorage.setItem(LOCAL_STORAGE_SELECTED_LIST_ID_KEY, selectedListId);
-}
-
-function render() {
-  clearElement(listsContainer);
-  renderLists();
-  const selectedList = lists.find((list) => list.id === selectedListId.value);
-  if (selectedListId.value == null) {
-    taskDisplayContainer.style.display = "none";
-  } else {
-    console.log("F render sl:", selectedList);
-    taskDisplayContainer.style.display = "block";
-    listTitleElement.innerText = selectedList.name;
-    renderTaskCount(selectedList);
-    clearElement(taskContainer);
-    renderTasks(selectedList);
-  }
-}
-
-// clearElement avoids duplication of previously displayed items
-function clearElement(element) {
-  while (element.firstChild) {
-    element.removeChild(element.firstChild);
-  }
-}
-
-function renderLists() {
-  lists.forEach((list) => {
-    const listElement = document.createElement("li");
-    listElement.dataset.listId = list.id;
-    listElement.classList.add("list-name");
-    listElement.innerText = list.name;
-    console.log("F rl sli:", selectedListId);
-    if (list.id === selectedListId.value) {
-      listElement.classList.add("active-list");
-    }
-    listsContainer.appendChild(listElement);
-  });
-}
-
-function renderTaskCount(selectedList) {
-  const incompleteTaskCount = selectedList.tasks.filter(
-    (task) => !task.complete
-  ).length;
-  const taskString = incompleteTaskCount === 1 ? "task" : "tasks";
-  taskCountElement.innerText = `${incompleteTaskCount} ${taskString} remaining`;
-}
-
-function renderTasks(selectedList) {
-  //Does this need to be broken down?
-  selectedList.tasks.forEach((task) => {
-    const taskElement = document.importNode(taskTemplate.content, true);
-    //console.log("F rTask,taskTemplate:", taskTemplate.content);
-    const checkbox = taskElement.querySelector("input");
-    checkbox.id = task.id;
-    checkbox.checked = task.complete;
-    const taskNameLabel = taskElement.querySelector(".task-name-label");
-    const taskNoteLabel = taskElement.querySelector(".task-note-label");
-    const taskPriorityLabel = taskElement.querySelector(".task-priority-label");
-    const taskDueDateLabel = taskElement.querySelector(".task-due-date-label");
-    taskNameLabel.htmlFor = task.id;
-    taskNoteLabel.htmlFor = task.id;
-    taskPriorityLabel.htmlFor = task.id;
-    taskDueDateLabel.htmlFor = task.id;
-    taskNameLabel.append(task.name);
-    taskNoteLabel.append(task.note);
-    taskPriorityLabel.append(task.priority);
-    taskDueDateLabel.append(task.dueDate);
-    taskContainer.appendChild(taskElement);
-  });
-}
+//  ^^^^^Functions^^^^^ migrated to functions2js
 
 render();
